+++ pulsar/apps/greenio/__init__.py
@@ -127,8 +127,7 @@ import logging
 from collections import deque
 from functools import wraps
 
-import greenlet
-from greenlet import getcurrent
+from greenlet import greenlet, getcurrent
 
 from pulsar import isfuture, async
 from pulsar import Future, get_event_loop, AsyncObject, is_async
@@ -142,14 +141,14 @@ class _DONE:
     pass
 
 
-class GreenletWorker(greenlet.greenlet):
+class GreenletWorker(greenlet):
     pass
 
 
 def wait(value):
     '''Wait for a possible asynchronous value to complete.
     '''
-    current = greenlet.getcurrent()
+    current = getcurrent()
     parent = current.parent
     return parent.switch(value) if parent else value
 
@@ -161,18 +160,18 @@ def run_in_greenlet(callable):
     '''
     @wraps(callable)
     def _(*args, **kwargs):
-        greenlet = GreenletWorker(callable)
+        green = GreenletWorker(callable)
         # switch to the new greenlet
-        result = greenlet.switch(*args, **kwargs)
+        result = green.switch(*args, **kwargs)
         # back to the parent
         while is_async(result):
             # keep on switching back to the greenlet if we get a Future
             try:
-                result = greenlet.switch((yield from result))
+                result = green.switch((yield from result))
             except Exception as exc:
-                result = greenlet.throw(exc)
+                result = green.throw(exc)
 
-        return greenlet.switch(result)
+        return green.switch(result)
 
     return _
 
@@ -185,17 +184,26 @@ class GreenPool(AsyncObject):
     '''
     worker_name = 'exec'
 
-    def __init__(self, max_workers=None, loop=None, maxtasks=None):
+    def __init__(self, max_workers=None, loop=None):
         self._loop = loop or get_event_loop()
         self._max_workers = min(max_workers or _DEFAULT_WORKERS, _MAX_WORKERS)
         self._greenlets = set()
         self._available = set()
-        self._maxtasks = maxtasks
         self._queue = deque()
         self._shutdown = False
         self._waiter = None
         self._logger = logging.getLogger('pulsar.greenpool')
         self._shutdown_lock = threading.Lock()
+
+    @property
+    def max_workers(self):
+        return self._max_workers
+
+    @max_workers.setter
+    def max_workers(self, value):
+        value = int(value)
+        assert value > 0
+        self._max_workers = value
 
     def submit(self, func, *args, **kwargs):
         '''Equivalent to ``func(*args, **kwargs)``.
@@ -225,10 +233,10 @@ class GreenPool(AsyncObject):
     def _adjust_greenlet_count(self):
         if (not self._shutdown and not self._available and
                 len(self._greenlets) < self._max_workers):
-            greenlet = GreenletWorker(self._green_run)
-            self._greenlets.add(greenlet)
+            green = GreenletWorker(self._green_run)
+            self._greenlets.add(green)
             self.logger.debug('Num greenlets: %d', len(self._greenlets))
-            greenlet.switch()
+            green.switch()
         return self._available
 
     def _put(self, task):
@@ -247,13 +255,13 @@ class GreenPool(AsyncObject):
             return
         async(self._green_task(self._available.pop(), task), loop=self._loop)
 
-    def _green_task(self, greenlet, task):
+    def _green_task(self, green, task):
         # Coroutine executing the in main greenlet
-        # This coroutine is executed for every task pu into the queue
+        # This coroutine is executed for every task put into the queue
 
         while task is not _DONE:
             # switch to the greenlet to start the task
-            task = greenlet.switch(task)
+            task = green.switch(task)
 
             # if an asynchronous result is returned, yield from
             while is_async(task):
@@ -261,21 +269,19 @@ class GreenPool(AsyncObject):
                     task = yield from task
                 except Exception as exc:
                     # This call can return an asynchronous component
-                    task = greenlet.throw(exc)
+                    task = green.throw(exc)
 
     def _green_run(self):
         # The run method of a worker greenlet
         task = True
         while task:
-            greenlet = getcurrent()
-            parent = greenlet.parent
+            green = getcurrent()
+            parent = green.parent
             assert parent
             # add greenlet in the available greenlets
-            self._available.add(greenlet)
+            self._available.add(green)
             task = parent.switch(_DONE)  # switch back to the main execution
             if task:
-                # If a new task is available execute it
-                # Here we are in the child greenlet
                 future, func, args, kwargs = task
                 try:
                     result = func(*args, **kwargs)
@@ -284,13 +290,80 @@ class GreenPool(AsyncObject):
                 else:
                     future.set_result(result)
             else:  # Greenlet cleanup
-                self._greenlets.remove(greenlet)
+                self._greenlets.remove(green)
                 if self._greenlets:
                     self._put(None)
                 elif self._waiter:
                     self._waiter.set_result(None)
                     self._waiter = None
                 parent.switch(_DONE)
+
+
+class GreenLock:
+    '''A Locking primitive that is owned by a particular greenlet
+    when locked.The main greenlet cannot acquire the lock.
+
+    A primitive lock is in one of two states, 'locked' or 'unlocked'.
+
+    It is created in the unlocked state. It has two basic methods,
+    :meth:`.acquire` and :meth:`.release. When the state is unlocked,
+    :meth:`.acquire` changes the state to locked and returns immediately.
+
+    When the state is locked, :meth:`.acquire` blocks the current greenlet
+    until a call to :meth:`.release` changes it to unlocked,
+    then the :meth:`.acquire` call resets it to locked and returns.
+    '''
+    def __init__(self, loop=None):
+        self._loop = loop or get_event_loop()
+        self._local = threading.local()
+        self._local.locked = None
+        self._queue = deque()
+
+    def locked(self):
+        ''''Return the greenlet that acquire the lock or None.
+        '''
+        return self._local.locked
+
+    def acquire(self, timeout=None):
+        '''Acquires the lock if in the unlocked state otherwise switch
+        back to the parent coroutine.
+        '''
+        green = getcurrent()
+        parent = green.parent
+        if parent is None:
+            raise RuntimeError('acquire in main greenlet')
+
+        if self._local.locked:
+            future = Future(loop=self._loop)
+            self._queue.append(future)
+            parent.switch(future)
+
+        self._local.locked = green
+        return self.locked()
+
+    def release(self):
+        '''Release the lock.
+
+        This method should only be called in the locked state;
+        it changes the state to unlocked and returns immediately.
+        If an attempt is made to release an unlocked lock,
+        a RuntimeError will be raised.
+        '''
+        if self._local.locked:
+            while self._queue:
+                future = self._queue.popleft()
+                if not future.done():
+                    return future.set_result(None)
+            self._local.locked = None
+        else:
+            raise RuntimeError('release unlocked lock')
+
+    def __enter__(self):
+        self.acquire()
+        return self
+
+    def __exit__(self, type, value, traceback):
+        self.release()
 
 
 class WsgiGreen: