+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Development
 * Added optional timeout to :class:`.HttpClient` requests.
 * Refactored :class:`.String` and renamed from ``AsyncString``. ``AsyncString``
   still available for backward compatibility.
+* Added the new :class:`.GreenLock` class, a locking primitive for
+  greenlets in a greenlet pool.
 * Added new example to snippets directory. A simple Actor application.
 
 Ver. 1.0.0 - 2015-May-18